+++ mayan/apps/common/exceptions.py
@@ -15,6 +15,19 @@ class ActionError(BaseCommonException):
     it is used to announce that one item in the queryset failed to process.
     """
 
+class CompressionFileError(BaseCommonException):
+    """
+    Base exception for file decompression class
+    """
+    pass
+
+
+class NoMIMETypeMatch(CompressionFileError):
+    """
+    There is no decompressor registered for the specified MIME type
+    """
+    pass
+
 
 class NotLatestVersion(BaseCommonException):
     """