+++ HISTORY.rst
@@ -127,6 +127,8 @@
 - Remove usage of pace.js. Would cause XMLRequest to fallback to
   synchronous mode.
 - Add custom AJAX spinner.
+- Complete refactor of the compress archive class support. Closes
+  GitLab issue #7.
 
 3.0.3 (2018-08-17)
 ==================